refactor: move followship function to followship controller, hide some sensitive information
